requirements.txt
pokebase
pyfiglet
requests
Pillow
